pyproject.toml
[project]
name = "query-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docling>=2.57.0",
    "fastapi>=0.116.1",
    "guardrails-ai>=0.6.7",
    "hf-xet>=1.1.9",
    "joblib>=1.5.2",
    "jsonref>=1.1.0",
    "langchain>=0.3.27",
    "langchain-core>=0.3.75",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.10",
    "langchain-huggingface>=0.3.1",
    "langgraph>=0.6.6",
    "langgraph-supervisor>=0.0.29",
    "llama-index>=0.14.0",
    "numpy>=2.3.2",
    "opentelemetry-api>=1.37.0",
    "opentelemetry-sdk>=1.37.0",
    "pandas>=2.2.3",
    "pinecone>=7.3.0",
    "pydash>=8.0.5",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "regex>=2025.9.1",
    "scikit-learn>=1.7.1",
    "sentence-transformers>=5.1.0",
    "uvicorn>=0.35.0",
]
